+++ meeting-appointment-modules/src/main/java/com/jfzt/meeting/service/MeetingDeviceService.java
@@ -8,8 +8,6 @@ import com.jfzt.meeting.entity.dto.MeetingDeviceDTO;
 import com.jfzt.meeting.entity.dto.MeetingDevicePageDTO;
 import com.jfzt.meeting.entity.vo.MeetingDeviceVO;
 
-import java.util.List;
-
 /**
  * 针对表【meeting_device(设备表)】的数据库操作Service
  * @author chenyu.di
@@ -22,42 +20,42 @@ public interface MeetingDeviceService extends IService<MeetingDevice> {
      * @param meetingDevicePageDTO 获取设备信息入参请求体
      * @return 设备信息
      */
-    Result<Page<MeetingDeviceVO>> getDevice(MeetingDevicePageDTO meetingDevicePageDTO);
+    Result<Page<MeetingDeviceVO>> getDevice (MeetingDevicePageDTO meetingDevicePageDTO);
 
     /**
      * 添加设备
      * @param meetingDeviceDTO 添加设备入参请求体
      * @return 成功信息
      */
-    Result<Object> addDevice(MeetingDeviceDTO meetingDeviceDTO);
+    Result<Object> addDevice (MeetingDeviceDTO meetingDeviceDTO);
 
     /**
      * 修改设备
      * @param meetingDevice 修改设备入参请求体
      * @return 成功信息
      */
-    Result<Object> updateDevice(MeetingDevice meetingDevice);
+    Result<Object> updateDevice (MeetingDevice meetingDevice);
 
     /**
      * 删除设备
      * @param id 设备id
      * @return 成功信息
      */
-    Result<Object> deleteDevice(Integer id);
+    Result<Object> deleteDevice (Long id);
 
     /**
      * 启用禁用
      * @param id 设备id
      * @return 成功信息
      */
-    Result<Object> updateStatusDevice(Long id);
+    Result<Object> updateStatusDevice (Long id);
 
     /**
      * 批量删除设备
      * @param ids 所选id集合
      * @return 删除结果
      */
-    Result<Object> deleteDevices(Integer[] ids);
+    Result<Object> deleteDevices (Long[] ids);
 
-    Result<Object> updateStopSend(Long id);
+    Result<Object> updateStopSend (Long id);
 }
